AES key generation
I was uncertain of how I was going to create the AES key, as it would normally be a text string to be shareable, but the encryption of the AES string would most definitely turn it from an ASCII character to an unreadable binary value. Therefore, I think I will make it so that the AES key is a bunch of numbers representing the byte instead of the byte itself. A byte has the max size of 256, so I will make the key a list of numbers [0, 256>, and the number of numbers like this will be 128/8, because the AES key has to be 128 bits long, and there are 8 bits in a byte, and a number is a whole byte in size.
I also changed all the files to txt files, such that i could specify that the filename parameter was the input, and the text files were the output later.